Created bme sensor controller
